+++ src/renderer/components/ui/Theme/Theme.css
@@ -1,12 +1,12 @@
 /* TODO these imports should be phased out */
-@import "../../styles/_colors.css";
-@import "../../styles/_sizes.css";
+@import "../../../styles/_colors.css";
+@import "../../../styles/_sizes.css";
 
 @font-face {
   font-family: "Open Sans";
   font-style: normal;
   font-weight: 300 400 500 600 700 800;
-  src: url("fonts/OpenSans.ttf") format("truetype");
+  src: url("./OpenSans.ttf") format("truetype");
 }
 
 body {
